backend/requirements.txt
backend/requirements.txt
flask
sentence-transformers
faiss-cpu
numpy
ollama
flask-cors